backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
setuptools>=68.0.0
aiohttp>=3.8.0
charset-normalizer==3.4.2
click==8.2.1
fastapi==0.115.12
h11==0.16.0
idna==3.10
moviepy==2.2.1
opencv-python==4.11.0.86
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
python-dotenv==1.1.0
requests==2.32.4
scenedetect==0.6.4
sniffio==1.3.1
tenacity==9.1.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
webrtcvad==2.0.10
yt-dlp==2025.6.9
# Subtitle subsystem dependencies
groq>=0.15.0
pytest==8.3.4
pytest-cov==6.0.0
pytest-asyncio==0.24.0
black==24.10.0
ffmpeg-python==0.2.0

sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9

python-jose[cryptography]==3.3.0
python-multipart==0.0.6

pydantic-settings==2.1.0

# JWT and Clerk authentication
PyJWT==2.8.0
cryptography==41.0.7

# CORS and security

# Azure Blb Storage
azure-storage-blob==12.19.0
azure-identity==1.15.0
aiofiles==23.2.0
